refactor(mc): rename mcmc.js to _mcmc.js and make MCMC abstract
- git mv mcmc.js _mcmc.js: base class gets the _ prefix convention,
  consistent with _distribution.js and _process.js
- Add new.target guard in constructor: throws if MCMC is instantiated
  directly rather than through a subclass
- Update import in rwm.js
